requirements.txt
pytest>=8.0.0
pytest-playwright>=0.4.0
pytest-base-url>=2.0.0
